changes to includes, formatting, wpm_render function, improved and simplified oled code, general cleanup.
