+++ keyboards/handwired/bongo80/rules.mk
@@ -1,0 +1,1 @@
+# This file intentionally left blank
